Don't use a global pointer to track loopback adapter
Using a function allows us to ensure the filter module (open instance)
is in the running state, preventing sync problems between the WFP
loopback stuff and the NDIS LWF stuff. Problems may still exist, but
this ought to help.

Potentially fixes a crash, access violation in NPF_NetworkClassify, that
seems to be result of trying to lock the loopback adapter's open
instance when it is not running.