Rearrange initialization of filter module list
As soon as we call NdisFRegisterFilterDriver, NDIS can start attaching
adapters. All members have to be initialized already. Additionally, the
global `g_arrFiltMod` has to be initialized when declared because
otherwise the compiler keeps rearranging things and setting `Next` to
NULL *after* the adapters are attached, resulting in NPF_OpenAdapter
always failing.