Clarify the systems affected by issue 1856
I don't think Windows 10 is the only system 0.9985 worked on--I think Windows Server 2018 and perhaps other Windows Server editions worked too.  Also, my understanding is that it still worked on older systems with driver signing turned off (admittedly that is a rare configuration change, but I think that's why we missed it in testing).