New mode, TIMESTAMPMODE_QUERYSYSTEMTIME_PRECISE
This mode uses KeQuerySystemTimePrecise if available. On Win7/2008R2, it
falls back to KeQuerySystemTime, equivalent to TimestampMode 2. While
this is not as precise as TIMESTAMPMODE_SINGLE_SYNCHRONIZATION, it at
least does not drift from wall clock time in the same way. I believe the
behavior (monotonic vs synchronized) is a more important distinction to
preserve than the precision.

Once we have per-capture-handle timestamp modes, we can probably have a
feedback mechanism to inform the caller whether a timestamp mode is
supported on the current platform.

See nmap/nmap#1407