New allocation scheme for larger packets.
For packets less than 256B, NDIS's NetBufferPool is efficient enough.
For larger ones, we need to allocate more buffer space and MDLs, and we
were freeing those as soon as we were done; very inefficient. Instead,
we'll put them on a stack and pull them back off when we need them. A
separate thread will clear the stack when all capture handles are
closed.
Related to #213	because by avoiding frequent small (257 to 1258 bytes)
allocations we reduce the chance of heap fragmentation. Also, this
changes pretty much all the code implicated in that issue, so if it
persists it's something new.