Fix race condition; new callbacks for ObjPool
There was a race condition based on the NPF_NB_COPIES::ulRefcount
hitting zero and the object being cached before another open instance in
TapExForEachOpen could increase that refcount, leading to duplicate
entries in the cache. Then one pointer would be in use while another one
got freed/returned, leading to a crash.

New design of ObjPool slab allocator allows callback functions to take a
context object as an argument, as well as allowing the CleanupFunc
callback to cancel a Return operation if it refcounts the object and
stashes it somewhere else (like the cache). This eliminates the race
condition.