Don't use a pointer after it's been returned.
The Curr pointer was part of the NBLCopy that was returned, so using it
again right afterwards could lead to traversing a different chain.
Defensive practice in next commit will be to zero memory when adding it
back into the ObjPool, ensuring this would be a null pointer deref
instead.