Remove driver's TimeOut member and update docs/comments
The BIOCSRTIMEOUT ioctl has not been used by Packet.dll since WinPcap
was ported to Windows NT. Instead, the timeout is only respected by
PacketReceivePacket in user mode. This change removes the IOCTL handler
and the TimeOut member as well as updating the comments in several files
to correctly describe the behavior of Packet.dll functions and the
ReadFile IRP handler.