Don't artificially limit writes by waiting until they finish.
NDIS 5 had a hard upper limit on the number of packets that could be
allocated from a packet pool (WinPcap sets this limit to 256). NDIS 6
NBL pools don't have such a limit, so we will just return appropriate
status if an allocation fails instead of trying to wait and work around
it. This will increase speed of network sends.