Track whether OID_GEN_CURRENT_PACKET_FILTER is allowed as query
WiFi adapters on Windows 10 and 11 return NDIS_STATUS_INVALID_OID for
this query. Previously, this was ignored and we would probably set an
empty packet filter on exit, but since we started checking return value,
it led to inability to capture at all, since libpcap expects to be able
to set the packet filter. No explanation so far for why WiFi adapters
are different in this way, but if any adapter returns
NDIS_STATUS_INVALID_OID, we will avoid sending it again in the future.