Refactor and redesign packet capture and related locks
To improve performance by reducing lock contention and hopefully address
issue #663, this change avoids holding a filter module's
OpenInstancesLock while processing packet data. Instead, a separate lock
is used to manage the list of currently active (OpenRunning) instances,
so that listing adapters or opening new instances does not use the same
locks as processing packets.

Another change in this commit is that each packet is matched against
each BPF filter before any packet data is copied, so that the copy
operation can be done once for the maximum required snaplen instead of
in stages if later filters match larger portions of the packet than
earlier ones. Furthermore, the BpfProgramsLock is released prior to
dispatching the packet data to each instance's queue, reducing the
amount of time holding the lock (at DISPATCH_LEVEL).