Fix race condition when closing a handle
Race condition has not been reported in any release build, but was
observed in debug build with Driver Verifier running. TOCTOU issue where
NPF_DoTap checks that the packet passes the Open's BPF filter, then
allows the Open to be closed before trying to dispatch the packet to the
Open via NPF_TapExForEachOpen. Solution is to mark the Open as in-use
before queueing the packet for dispatch, then release it after dispatch.