Eliminate most copies during sends. Support VLAN tags in BufferedWrite.
There's no reason to copy most packet data during a send operation as
long as we ensure the buffer is not going to page out before the send is
complete. Previous page fault crashes were due to one of a few issues:

1. Old ioctl codes (pre-0.9981) used methods other than METHOD_BUFFERED,
   so the buffer may not have been probed-and-locked.
2. We may have completed the IRP before the send was complete, so the
   I/O manager called MmUnlockPages.

The solution is to refcount the send and only complete the IRP when the
send is completed. We already did this for NPF_BufferedWrite. NPF_Write
benefits from refcount on the IRP itself in case sends are reordered.